WIP: move peak_finder
