+++ mne/preprocessing/__init__.py
@@ -15,6 +15,7 @@ from .ecg import find_ecg_events, create_ecg_epochs
 from .ica import (ICA, ica_find_eog_events, ica_find_ecg_events,
                   get_score_funcs, read_ica, run_ica, corrmap)
 from .otp import oversampled_temporal_projection
+from . import peak_finder
 from .bads import find_outliers
 from .infomax_ import infomax
 from .stim import fix_stim_artifact
